test: adjust coordinator test formatting
