--- src/testing/config.py
@@ -1,5 +1,0 @@
-"""Shared configuration and constants for the testing orchestrator."""
-from __future__ import annotations
-
-TEST_FILE_PATTERN = "test_*.py"
-COVERAGE_REPORT_PRECISION = 2
