+++ testing/__init__.py
@@ -1,0 +1,5 @@
+"""Testing utilities coordinating setup, execution, and teardown."""
+
+from .coordinator import TestCoordinator
+
+__all__ = ["TestCoordinator"]
