+++ testing/infrastructure/config.py
@@ -1,0 +1,4 @@
+"""Shared configuration for test infrastructure."""
+
+TEST_FILE_PATTERN = "test_*.py"
+COVERAGE_REPORT_PRECISION = 2
